mwe for bernoulli posterior in tensorflow

+++ blackbox/dists.py
@@ -13,8 +13,8 @@ def bernoulli_log_prob(x, p):
         # TODO: adapt to float64, etc.
         p = tf.clip_by_value(p, 1e-45, 1.0)
 
-    lp = tf.log(p)
-    lp1 = tf.log(1.0-p)
+    lp = tf.to_float(tf.log(p))
+    lp1 = tf.to_float(tf.log(1.0-p))
     return tf.mul(x, lp) + tf.mul(1-x, lp1)
 
 def beta_log_prob(x, alpha=1.0, beta=1.0):
